modified:   demo/src/main/java/com/example/List.java 	modified:   demo/src/main/java/com/example/Main.java

+++ demo/src/main/java/com/example/List.java
@@ -92,7 +92,7 @@ public class List {
 
 	public int getElem(int ind){
 		if(ind >= size){
-			return null;
+			return (Integer) null;
 		}	
 		else{
 			Node ptr = head;
